+++ src/jarvis/jarvis.module.ts
@@ -44,6 +44,9 @@ import { CategorySummaryService } from './library/category-summary.service';
 import { DocumentSummaryService } from './library/document-summary.service';
 import { DocumentCompareService } from './library/document-compare.service';
 import { KnowledgeTestService } from './library/knowledge-test.service';
+import { GoogleGmailService } from './tools/google/google-gmail.service';
+import { GoogleDriveService } from './tools/google/google-drive.service';
+import { YouTubeService } from './tools/google/youtube.service';
 
 @Module({
   imports: [PrismaModule, GoogleModule],
@@ -67,6 +70,9 @@ import { KnowledgeTestService } from './library/knowledge-test.service';
     JarvisIdentityService,
     GoogleCalendarService,
     GoogleTasksService,
+    GoogleGmailService,
+    GoogleDriveService,
+    YouTubeService,
     AstrologyTool,
     MemoryExtractorService,
     CapabilitiesService,